update schematic to use fets, txs0108 not suitable (+)
the txs0108 doesn't do well with capacitive loads, so the reset line
keeps dropping too low (around 3.8v or below); the knock-on effect is
that the amiga keeps resetting itself but moreover controller ports do
not work correctly (if at all), and the keyboard behaves erratically
(though the keycodes get through, the voltage levels are too low to
match spec).

tried with txb0108; effect was similar, but actually worse in many ways.

the first production board revision will use FETs, but this may change
if i find a more suitable logic level converter - i don't particularly
want to give up the bidirectional nature of many of the lines because
one goal right now is to use some of the bidirectional lines for serial
data transfer to change config settings on the hid-pico board from the
amiga side.