don't raise output; set dir=in and let amiga do it
cryptic short message, but by not raising the output to high level we're
not actively sending a high signal, instead relying on the amiga doing
so. ultimately, with the current mosfet-based design, it doesn't matter
a great deal: either side of the voltage conversion uses a pullup, so a
current is raised regardless, but the pico's drain to 0V (active low)
will be unaffected.